added composer autoloading and deps management
