+++ .gitignore
@@ -1,1 +1,3 @@
 *.swp
+.idea
+/vendor/
