+++ src/AuthException.php
@@ -1,0 +1,9 @@
+<?php 
+namespace TidalLabs\Slacker;
+use Exception;
+
+
+/**
+ * Used for Auth or token errors.
+ */
+class AuthException extends Exception { }
